Renamed signupSlice with authSlice.
